pyproject.toml
[tool.poetry]
name = "coworkings_be"
version = "0.1.0"
description = ""
authors = ["Aliaksei Valenta <liverpoolsmile23@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.2"
psycopg2-binary = "^2.9.4"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.4"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
